Finding the best buy and sell prices for commodities
At this point the build script spits out the 20 best commodity trades in
the game including where to buy and where to sell each commodity.

Next piece I'd like to tackle is have it tell you what the best
commodity to buy in one system is and where to sell it. Would also like
to write something to tell you the best trade between two specific
points is.