Discovered the Flint API (fl-flint)
A couple people on the DiscoveryGC discord pointed me to the flint api
(https://github.com/biqqles/flint). It makes getting information out of
the game files much easier and more accurate. It however does not seem
to be able to take into account price overrides from plugins, so that
part stayed in.