Accounting for mismatched base ids
Planet Sprague in Omega-3 has the id of br04_01_base in the game files
when it should be bw01_01_base.

Planet New Berlin in New Berlin has the id of rh10_02_base in the game
files when it should be rh01_01_base.

The "correct" names are referenced in the prices.cfg override plugin
file, but I'm not sure if in game it is applied correclty since those
bases have different ids. Setting them correctly will at least get the
right original prices to the right base for our script. Otherwise either
the misnamed base or the duplicated base (whichever is processed first)
would end up being missing in the price lists.